requirements.txt
aiohttp==3.7.4.post0
antlr4-python3-runtime==4.9.2
anyio==3.1.0
argon2-cffi==20.1.0
async-generator==1.10
async-timeout==3.0.1
attrs==21.2.0
autobahn==21.3.1
backcall==0.2.0
bleach==3.3.0
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
colorama==0.4.4
cryptography==3.4.7
decorator==5.0.9
defusedxml==0.7.1
entrypoints==0.3
hyperlink==21.0.0
idna==3.2
ipykernel==5.5.5
ipython==7.24.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
jedi==0.18.0
Jinja2==3.0.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.12
jupyter-console==6.4.0
jupyter-core==4.7.1
jupyter-server==1.8.0
jupyter-server-proxy==3.0.2
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
MarkupSafe==2.0.1
matplotlib-inline==0.1.2
mistune==0.8.4
multidict==5.1.0
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.1
notebook==6.4.0
numpy==1.20.3
packaging==20.9
pandocfilters==1.4.3
parso==0.8.2
pickleshare==0.7.5
prometheus-client==0.10.1
prompt-toolkit==3.0.18
pycparser==2.20
Pygments==2.9.0
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
pywin32==300
pywinpty==1.1.1
pyzmq==22.1.0
qtconsole==5.1.0
QtPy==1.9.0
Send2Trash==1.5.0
simpervisor==0.4
six==1.16.0
sniffio==1.2.0
terminado==0.10.0
testpath==0.5.0
tornado==6.1
traitlets==5.0.5
txaio==21.2.1
typing-extensions==3.10.0.0
vpython==7.6.1
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.0.1
widgetsnbextension==3.5.1
wincertstore==0.2
yarl==1.6.3
